Restructured fields tests and prepped additional fields
