--- src/tests/when_no_animal_information_is_provided.cs
@@ -1,31 +1,0 @@
-using System;
-using System.Threading;
-using Moq;
-using PetHome.Handlers;
-using PetHome.Handlers.Commands;
-using PetHome.Persistence;
-using Xunit;
-
-namespace tests
-{
-    public class when_no_animal_information_is_provided
-    {
-        private RegisterNewAnimalHandler Subject;
-        private Mock<IApplicationContext> Context;
-
-        public when_no_animal_information_is_provided()
-        {
-            Context = new Mock<IApplicationContext>();
-
-            Subject = new RegisterNewAnimalHandler(Context.Object);
-        }
-
-        [Fact]
-        public void argument_exception_is_thrown()
-        {
-            Action action = () => { Subject.Handle(new RegisterNewAnimal() { }, CancellationToken.None); };
-
-            Assert.Throws<ArgumentException>(action);
-        }
-    }
-}
